pyproject.toml
[tool.poetry]
name = "img_to_pdf"
version = "0.1.0"
description = ""
authors = ["Hemang Kumar <hemangsk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.25.0"
pillow = "^10.4.0"
reportlab = "^4.2.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
